Lots of changes big and small
- Changed PColor to use floats instead of doubles for color components.
Floats make more sense here since all major render bindings use floats
rather than doubles. Implementations have been changed accordingly.
- Added ImmutablePColorInt32 implementation of PColor for memory
efficient storage of color.
- Added AwtPColor implementation of PColor for Swing root implementation
- Replaced the component tree iteration methods in PComponent with the
new ComponentStream API. Still WIP to a certain extend.
- Made all PSelectionComponents to PInteractiveComponents. Fixed
implementation of PList, PTable and PTree accordingly.
- Introduced an interface PListLike for components with a similar
functionality as a PList. This will be used soon for an alternative
PList implementation.
- Fixed raw usage of single value model and observer in PTitleBorder
- Refactored the classes within the scroll2 package and fixed some bugs.
Still a WIP.
- Added convenient methods to PBounds to interface with the PPoint API
- Added convenient methods to PPoint. Fixed bug in AbstractPPoint
implementation of equals
- Made the ZERO_INSETS in PInsets an instance of AbstractPInsets. This
will provide compatibility of the hashCode and equals methods.
- Changed accessibility of attributes in MutablePSize to protected
- Wrote better java docs for several interfaces, classes and methods.
Still WIP.
- Fixed lots of existing documentation
- Removed old / dead code

+++ piq-core/src/edu/udo/piq/PColor.java
@@ -12,7 +12,7 @@ import edu.udo.piq.tools.ImmutablePColor;
  * 
  * @see PRenderer
  * @see PRenderer#setColor(PColor)
- * @see PRenderer#setColor1(double, double, double, double)
+ * @see PRenderer#setColor1(float, float, float, float)
  * @see PRenderer#setColor255(int, int, int, int)
  */
 public interface PColor {
@@ -20,70 +20,70 @@ public interface PColor {
 	/**
 	 * An immutable PColor implementation representing absolute black with full transparency (alpha == 0).
 	 */
-	public static final PColor TRANSPARENT	= new ImmutablePColor(0.00, 0.00, 0.00, 0.00);
+	public static final PColor TRANSPARENT	= new ImmutablePColor(0.00f, 0.00f, 0.00f, 0.00f);
 	/**
 	 * An immutable PColor implementation representing absolute white.
 	 */
-	public static final PColor WHITE	= new ImmutablePColor(1.00, 1.00, 1.00);
+	public static final PColor WHITE	= new ImmutablePColor(1.00f, 1.00f, 1.00f);
 	/**
 	 * An immutable PColor implementation representing a grey tone of about 25%.
 	 */
-	public static final PColor GREY25	= new ImmutablePColor(0.25, 0.25, 0.25);
+	public static final PColor GREY25	= new ImmutablePColor(0.25f, 0.25f, 0.25f);
 	/**
 	 * An immutable PColor implementation representing a grey tone of about 50%.
 	 */
-	public static final PColor GREY50	= new ImmutablePColor(0.50, 0.50, 0.50);
+	public static final PColor GREY50	= new ImmutablePColor(0.50f, 0.50f, 0.50f);
 	/**
 	 * An immutable PColor implementation representing a grey tone of about 75%.
 	 */
-	public static final PColor GREY75	= new ImmutablePColor(0.75, 0.75, 0.75);
+	public static final PColor GREY75	= new ImmutablePColor(0.75f, 0.75f, 0.75f);
 	/**
 	 * An immutable PColor implementation representing a grey tone of about 87.5%.
 	 */
-	public static final PColor GREY875	= new ImmutablePColor(0.875, 0.875, 0.875);
+	public static final PColor GREY875	= new ImmutablePColor(0.875f, 0.875f, 0.875f);
 	/**
 	 * An immutable PColor implementation representing absolute black.
 	 */
-	public static final PColor BLACK	= new ImmutablePColor(0.00, 0.00, 0.00);
+	public static final PColor BLACK	= new ImmutablePColor(0.00f, 0.00f, 0.00f);
 	/**
 	 * An immutable PColor implementation representing absolute red.
 	 */
-	public static final PColor RED		= new ImmutablePColor(1.00, 0.00, 0.00);
-	public static final PColor DARK_RED		= RED.mult1(0.5);
-	public static final PColor LIGHT_RED	= RED.add1(0.5);
+	public static final PColor RED		= new ImmutablePColor(1.00f, 0.00f, 0.00f);
+	public static final PColor DARK_RED		= RED.mult1(0.5f);
+	public static final PColor LIGHT_RED	= RED.add1(0.5f);
 	/**
 	 * An immutable PColor implementation representing absolute green.
 	 */
-	public static final PColor GREEN	= new ImmutablePColor(0.00, 1.00, 0.00);
-	public static final PColor DARK_GREEN	= GREEN.mult1(0.5);
-	public static final PColor LIGHT_GREEN	= GREEN.add1(0.5);
+	public static final PColor GREEN	= new ImmutablePColor(0.00f, 1.00f, 0.00f);
+	public static final PColor DARK_GREEN	= GREEN.mult1(0.5f);
+	public static final PColor LIGHT_GREEN	= GREEN.add1(0.5f);
 	/**
 	 * An immutable PColor implementation representing absolute blue.
 	 */
-	public static final PColor BLUE		= new ImmutablePColor(0.00, 0.00, 1.00);
-	public static final PColor DARK_BLUE	= BLUE.mult1(0.5);
-	public static final PColor LIGHT_BLUE	= BLUE.add1(0.5);
+	public static final PColor BLUE		= new ImmutablePColor(0.00f, 0.00f, 1.00f);
+	public static final PColor DARK_BLUE	= BLUE.mult1(0.5f);
+	public static final PColor LIGHT_BLUE	= BLUE.add1(0.5f);
 	/**
 	 * An immutable PColor implementation representing absolute magenta
 	 * (red = 100%, blue = 100%).
 	 */
-	public static final PColor MAGENTA	= new ImmutablePColor(1.00, 0.00, 1.00);
-	public static final PColor DARK_MAGENTA		= MAGENTA.mult1(0.5);
-	public static final PColor LIGHT_MAGENTA	= MAGENTA.add1(0.5);
+	public static final PColor MAGENTA	= new ImmutablePColor(1.00f, 0.00f, 1.00f);
+	public static final PColor DARK_MAGENTA		= MAGENTA.mult1(0.5f);
+	public static final PColor LIGHT_MAGENTA	= MAGENTA.add1(0.5f);
 	/**
 	 * An immutable PColor implementation representing absolute yellow
 	 * (red = 100%, green = 100%).
 	 */
-	public static final PColor YELLOW	= new ImmutablePColor(1.00, 1.00, 0.00);
-	public static final PColor DARK_YELLOW	= YELLOW.mult1(0.5);
-	public static final PColor LIGHT_YELLOW	= YELLOW.add1(0.5);
+	public static final PColor YELLOW	= new ImmutablePColor(1.00f, 1.00f, 0.00f);
+	public static final PColor DARK_YELLOW	= YELLOW.mult1(0.5f);
+	public static final PColor LIGHT_YELLOW	= YELLOW.add1(0.5f);
 	/**
 	 * An immutable PColor implementation representing absolute teal
 	 * (green = 100%, blue = 100%).
 	 */
-	public static final PColor TEAL		= new ImmutablePColor(0.00, 1.00, 1.00);
-	public static final PColor DARK_TEAL	= TEAL.mult1(0.5);
-	public static final PColor LIGHT_TEAL	= TEAL.add1(0.5);
+	public static final PColor TEAL		= new ImmutablePColor(0.00f, 1.00f, 1.00f);
+	public static final PColor DARK_TEAL	= TEAL.mult1(0.5f);
+	public static final PColor LIGHT_TEAL	= TEAL.add1(0.5f);
 	
 	/**
 	 * Returns the value for the red color component in integer format.<br>
@@ -104,7 +104,7 @@ public interface PColor {
 	 * 
 	 * @return red color component between 0.0 and 1.0 inclusive.
 	 */
-	public double getRed1();
+	public float getRed1();
 	
 	/**
 	 * Returns the value for the green color component in integer format.<br>
@@ -125,7 +125,7 @@ public interface PColor {
 	 * 
 	 * @return green color component between 0.0 and 1.0 inclusive.
 	 */
-	public double getGreen1();
+	public float getGreen1();
 	
 	/**
 	 * Returns the value for the blue color component in integer format.<br>
@@ -146,7 +146,7 @@ public interface PColor {
 	 * 
 	 * @return blue color component between 0.0 and 1.0 inclusive.
 	 */
-	public double getBlue1();
+	public float getBlue1();
 	
 	/**
 	 * Returns the value for the alpha color component in integer format.<br>
@@ -167,7 +167,7 @@ public interface PColor {
 	 * 
 	 * @return alpha color component between 0.0 and 1.0 inclusive.
 	 */
-	public double getAlpha1();
+	public float getAlpha1();
 	
 	public default PColor add(PColor other) {
 		return add1(other.getRed1(), other.getGreen1(), other.getBlue1(), other.getAlpha1());
@@ -177,18 +177,18 @@ public interface PColor {
 		return mult1(other.getRed1(), other.getGreen1(), other.getBlue1(), other.getAlpha1());
 	}
 	
-	public default PColor add1(double increment) {
+	public default PColor add1(float increment) {
 		return add1(increment, increment, increment);
 	}
 	
-	public default PColor add1(double r, double g, double b) {
+	public default PColor add1(float r, float g, float b) {
 		r = PColor.addColorComponent1(getRed1(), r);
 		g = PColor.addColorComponent1(getGreen1(), g);
 		b = PColor.addColorComponent1(getBlue1(), b);
 		return new ImmutablePColor(r, g, b, getAlpha1());
 	}
 	
-	public default PColor add1(double r, double g, double b, double a) {
+	public default PColor add1(float r, float g, float b, float a) {
 		r = PColor.addColorComponent1(getRed1(), r);
 		g = PColor.addColorComponent1(getGreen1(), g);
 		b = PColor.addColorComponent1(getBlue1(), b);
@@ -196,18 +196,18 @@ public interface PColor {
 		return new ImmutablePColor(r, g, b, a);
 	}
 	
-	public default PColor mult1(double factor) {
+	public default PColor mult1(float factor) {
 		return mult1(factor, factor, factor);
 	}
 	
-	public default PColor mult1(double r, double g, double b) {
+	public default PColor mult1(float r, float g, float b) {
 		r = PColor.multColorComponent1(getRed1(), r);
 		g = PColor.multColorComponent1(getGreen1(), g);
 		b = PColor.multColorComponent1(getBlue1(), b);
 		return new ImmutablePColor(r, g, b, getAlpha1());
 	}
 	
-	public default PColor mult1(double r, double g, double b, double a) {
+	public default PColor mult1(float r, float g, float b, float a) {
 		r = PColor.multColorComponent1(getRed1(), r);
 		g = PColor.multColorComponent1(getGreen1(), g);
 		b = PColor.multColorComponent1(getBlue1(), b);
@@ -253,8 +253,8 @@ public interface PColor {
 		return mult255(factor, factor, factor);
 	}
 	
-	public static double addColorComponent1(double c1, double c2) {
-		double result = c1 + c2;
+	public static float addColorComponent1(float c1, float c2) {
+		float result = c1 + c2;
 		if (result < 0) {
 			return 0;
 		}
@@ -264,8 +264,8 @@ public interface PColor {
 		return result;
 	}
 	
-	public static double multColorComponent1(double c1, double c2) {
-		double result = c1 * c2;
+	public static float multColorComponent1(float c1, float c2) {
+		float result = c1 * c2;
 		if (result < 0) {
 			return 0;
 		}
@@ -287,7 +287,7 @@ public interface PColor {
 	}
 	
 	public static int multColorComponent255(int c1, int c2) {
-		return (int) (0.5 + 255.0 * PColor.multColorComponent1(c1 / 255.0, c2 / 255.0));
+		return (int) (0.5f + 255.0f * PColor.multColorComponent1(c1 / 255.0f, c2 / 255.0f));
 	}
 	
 }
